Address: Addresses shall relate to another
Addresses can not only be split into a tree
but also refer to some totally different addresses.
The difference between a sub-address
and a related address is,
that the related does not have to be a child or parent of the current address.
It can come from a total different place within the whole bulk of addresses.